server/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
python-multipart==0.0.9
python-dotenv==1.0.1
httpx==0.27.0
oss2==2.18.0
alibabacloud-imageenhan20190930==3.0.3
alibabacloud-imageseg20191230==3.0.2
alibabacloud-facebody20191230==5.0.2
alibabacloud-tea-openapi==0.3.8
Pillow==10.4.0
